+++ libs/btp/CHANGELOG.md
@@ -1,3 +1,32 @@
+## 0.62.3-rc.0 (2026-05-13)
+
+### 🚀 Features
+
+- MCP server evaluation - part 1 ([#14178](https://github.com/SAP/fundamental-ngx/pull/14178), [#14159](https://github.com/SAP/fundamental-ngx/issues/14159))
+
+### ❤️ Thank You
+
+- Maria Dineva @MariaIDineva
+
+## 0.62.2 (2026-05-13)
+
+### 🚀 Features
+
+- **core:** improve MCP server metadata and add usage guide tool ([#14159](https://github.com/SAP/fundamental-ngx/pull/14159))
+
+### 🩹 Fixes
+
+- **btp:** center vertically the Side Nav icons with nav button in Overlay Mode ([#14197](https://github.com/SAP/fundamental-ngx/pull/14197))
+
+### ❤️ Thank You
+
+- deno
+- Inna Atanasova @InnaAtanasova
+
+## 0.62.2-rc.17 (2026-05-12)
+
+This was a version bump only for btp to align it with other projects, there were no code changes.
+
 ## 0.62.2-rc.16 (2026-05-12)
 
 This was a version bump only for btp to align it with other projects, there were no code changes.